Move libs/ to lib/
