+++ lib/providers/node.js
@@ -8,7 +8,7 @@ module.exports = function(filePath) {
 	if (pkg.version === null || pkg.private) {
 		var manifest = 'manifest.json';
 		if (!test('-e', manifest)) {
-		      throw new Error("Null version or private flag detected and manifest.json not found");
+		  throw new Error("Null version or private flag detected and manifest.json not found");
 		}
 		console.log("Null version or private flag detected, switching to manifest.json");
 		this.filePath = manifest;
