+++ package.json
@@ -2,7 +2,7 @@
   "name": "releasy",
   "version": "1.5.6",
   "description": "CLI tool to release node applications with tag and auto semver bump",
-  "main": "libs/releasy.js",
+  "main": "lib/releasy.js",
   "bin": "bin/releasy.js",
   "scripts": {
     "test": "node_modules/.bin/mocha --compilers coffee:coffee-script/register"
